feat(simulation): multiple changes, see commit body for details
- refactor(models): convert every Amount and Balance from int to float64
- feat(simulation): improve readibility of console logs
- fix(simulation): disable payments from borrower to lenders when the loan is failed

+++ models/actor.go
@@ -10,6 +10,6 @@ type ModelBase interface {
 
 type Actor interface {
 	ModelBase
-	UpdateBalance(amount int)
+	UpdateBalance(amount float64)
 	GetID() uint
 }
